+++ src/components/MyInfo/Communinty/CreateInfo/Createlnfo.tsx
@@ -73,6 +73,5 @@ const CreateInfo = ({ userCreateCommunityData }: CreateInfoProps) => {
     </StArticle>
   );
 };
-console.log('userCreateCommunityData');
 
 export default CreateInfo;
